requirements.txt
agno
yfinance
python-dotenv
duckduckgo-search
brave-search
ddgs # duckduckgo-search
groq
streamlit